Add mkRun, mkUses
Take a _little_ bit of the burden of having to know whether to use run or
uses+with off of the caller.

I'm sort of wondering if _not_ exporting `Step.default` might be a good
thing, to discourage use of it without `mkRun` or `mkUses`. We'll see
over time, I think.